# src/docs/swagger.yaml
definitions:
...
  models.BasicErrorRes:
    properties:
      error:
        type: string
    type: object
...
  models.CreateService:
    properties:
      price:
        type: number
      serviceDescription:
        type: string
      serviceName:
        type: string
      serviceType:
        type: string
    type: object
...
  models.SearchFilter:
    properties:
      descending:
        type: boolean
      end_price_range:
        type: number
      end_time:
        type: string
      location:
        type: string
      max_rating:
        type: number
      min_rating:
        type: number
      page_number:
        type: integer
      page_size:
        type: integer
      q:
        type: string
      sort_by:
        description: price, rating, name
        type: string
      start_price_range:
        type: number
      start_time:
        type: string
    type: object
  models.SearchHistory:
    properties:
      search_date:
        type: string
      search_filters:
        $ref: '#/definitions/models.SearchFilter'
    type: object
  models.Service:
    properties:
      averageRating:
        type: number
      price:
        type: number
      requireCert:
        type: boolean
      serviceDescription:
        type: string
      serviceID:
        type: string
      serviceImg:
        items:
          type: integer
        type: array
      serviceName:
        type: string
      serviceType:
        type: string
      timeslots:
        items:
          $ref: '#/definitions/models.Timeslot'
        type: array
    type: object
  models.Timeslot:
    properties:
      endTime:
        type: string
      startTime:
        type: string
      status:
        type: string
      timeslotID:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /service/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Delete a service
      tags:
      - Service
    patch:
      consumes:
      - application/json
      description: Update a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Update a service
      tags:
      - Service
...
  /service/create:
    post:
      consumes:
      - application/json
      description: Create a new service
      parameters:
      - description: Service data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Create a service
      tags:
      - Service
  /service/duplicate/{id}:
    post:
      consumes:
      - application/json
      description: Duplicate a service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Service'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Duplicate a service
      tags:
      - Service
...
  /service/searching:
    post:
      consumes:
      - application/json
      description: Search services based on query, location, timeslot, price range,
        rating
      parameters:
      - description: Search history
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SearchHistory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Service'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      summary: Search services
      tags:
      - Service
...
swagger: "2.0"